feat(inventory): widen the inventory quantity chain to decimal, gated by the declared precision_scale
Stage 1 (#1413) made the whole-quantity invariant per-product without changing
a single type: a work-order part quantity must be integral unless the product's
BASE unit-of-measure row declares precision_scale > 0. This stage changes the
types so that declaring a non-zero scale actually means something end to end.

The ledger is the part #1365's Option A framing left out, and it is the reason
this could not stop at the read model. inventory_ledger_entry.change_in_quantity
and quantity_after, the reservation / allocation / backorder chain, the
inventory_stock_summary read model, the costing state, cycle-count variances,
returns, manual adjustments and shortage records all become numeric(19,4) /
BigDecimal — the same width workorder_part.quantity has always had.

Widening the columns does not make stock divisible. A product declaring scale 0,
which is every product until one is seeded otherwise, is still refused a
fractional quantity; the width exists so a product declaring scale > 0 has
somewhere truthful to land. UomConversionService.precisionScale finally has a
production caller, through a new declaredBaseScale that answers 0 for a product
that declares nothing.

Enforcement is symmetric, which is what resolves the contradiction that was
latent in main. UomConversionServiceImplTest carries a base-LB-at-scale-2
product and asserts 1 BAG -> 1.01 LB; feed that same product through receiving
and toWholeLedgerQuantity's intValueExact() threw. The conversion subsystem and
the ledger could not both be right about it. The three intValueExact() guards in
receiving, ASN and returns are re-driven off the declaration rather than
deleted: they keep their fail-closed character and lose their hardcoding, and
they now raise 422 FRACTIONAL_QUANTITY_NOT_ALLOWED — the same code stage 1's
demand-side gate raises, from the same catalog declaration. Manual stock
movements, which post to the ledger just as directly, gain the same gate. A
receiving posting of 1.01 for that scale-2 product now round-trips into the
ledger, pinned as a test.

On the read side the Math.toIntExact calls in InventoryAvailabilityServiceImpl
and InventoryFactPublisher threw rather than wrapping, and that fail-loud
property is the one worth keeping; the 32-bit bound they checked is not, because
a decimal read model has no edge to fall off. They are replaced by the check
that still means something after the widening: does this quantity carry more
precision than the product ever declared it could? A stock reference that names
no catalog product is reported as stored rather than refused — an absent
declaration cannot be evidence that a value already in the ledger is wrong, and
refusing would turn an availability read into a 500 and stall the replicas
behind it.

Stage 1 left WorkorderPartUsageServiceImpl.reservableQuantity throwing on a
non-integral quantity, correctly noting the throw was unreachable and would stop
being so here. The seam it guarded is gone: the reservation command and its
outcome fact are decimal, so the quantity passes through unchanged and the test
pinning the unreachable path is deleted. The command's idempotency key now
normalises the quantity with stripTrailingZeros, so a retry that serialises 2 as
2.00 still collapses to one command.

Deliberately still integral, and why: pick-task, putaway and transfer-order line
quantities, and the cycle-count capture DTOs, widen with the unit-of-measure
work in stages 3 and 4; where they meet the ledger they widen at the boundary
rather than narrowing it. Sales-order line quantities stay int per ADR-0055.
Serial-unit postings keep intValueExact because a serialised unit is discrete by
definition — that is a property of the tracking mode, not of the type. The
replenishment need rounds UP to whole orderable units, because a vendor and a
shipping label do not understand 3.5 of something and under-ordering leaves the
shortage unresolved.

The v2 forecast fields on InventoryAvailabilityUpdatedV1 were long alongside the
older ints. They widen too: projectedAvailable is onHand + incoming - outgoing,
so leaving them integral once on-hand is decimal would only relocate the
truncation into the projection, where it is harder to see. The DOWN floor on
expected incoming supply went with them — it existed to squeeze a fractional
quantity into a long, and there is no longer anything to squeeze it into.

Schema-version handling, per the AC. InventoryAvailabilityUpdatedV1,
ReservationOutcomeV1, BackorderCreatedV1, BackorderResolvedV1,
ProductValueChangedV1 and StorageLocationOnHandUpdatedV1 change their quantity
fields from int/long to BigDecimal. That is a breaking payload change for the
ext_inventory_availability replicas in pos-order, pos-workorder and pos-catalog
and for pos-location's ext_storage_location_on_hand replica. No dual-read shim
is added, deliberately: the platform is pre-production with no live data and
every producer and consumer of these topics ships from this one reactor, so
there is no version skew to bridge — only a deployment ordering to respect. The
constraint is stated rather than mitigated, in the pos-inventory README: deploy
the fleet together, run the widening migrations first, and expect an old
consumer to reject a new payload as a databind failure (loud, counted on
replica.payload.rejected) rather than land a wrong number. The migrations are
plain ALTER ... TYPE widenings; numeric(19,4) holds every value the integer
columns could, so the implicit cast preserves existing rows and no
data-preservation logic is needed.

pos-catalog is the module everybody forgets — outside the #1315 gate work, and
its existing listener test hand-writes the payload JSON, which would keep
passing through a type change on the record. It gains a test that serialises the
real contract type with fractional quantities, so a narrowing breaks there
rather than silently truncating a bulk product's on-hand on a product-detail
page.

pos-location was not in the issue's list of three replicas but consumes
StorageLocationOnHandUpdatedV1, whose only remaining narrowing site was the one
Math.toIntExact left in InventoryFactPublisher. Its "no decommission while
stocked" guard has to see a partial drum as stock rather than as zero, so it is
widened with the rest.

Bulk shrinkage tolerance (#1365's open question) is settled and is stage 4
(#1416): a physical count and an inventory adjustment stay separate
transactions, reconciled against a tolerance configurable per product and
storage location. Nothing here builds that workflow; what it does is make sure
the types can carry the decimal variance it will need, and cycle-count
adjustments widen for exactly that reason — an integral variance column would
round away the small fractional shrinkage a bulk SKU produces, which is the one
thing the platform gets to notice about a drum that is quietly draining.

Refs #1414, ADR-0055

+++ pos-catalog/src/main/java/com/positivity/catalog/internal/dto/ProductDetailView.java
@@ -4,6 +4,7 @@ import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.NOT_REQUIR
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.util.List;
@@ -108,10 +109,10 @@ public class ProductDetailView {
     @Schema(description = "Availability information with status indicator")
     public static class AvailabilityInfo {
         @Schema(description = "On-hand quantity at location", example = "15", requiredMode = NOT_REQUIRED)
-        private Integer onHandQuantity;
+        private BigDecimal onHandQuantity;
 
         @Schema(description = "Available to promise quantity", example = "12", requiredMode = NOT_REQUIRED)
-        private Integer availableToPromiseQuantity;
+        private BigDecimal availableToPromiseQuantity;
 
         @Schema(description = "Lead time information", requiredMode = NOT_REQUIRED)
         private LeadTimeInfo leadTime;
